moved data fetch function ot its own file, enabled cloudflare image loader.
