+++ next.config.ts
@@ -3,8 +3,8 @@ import type { NextConfig } from "next";
 const nextConfig: NextConfig = {
   /* config options here */
   images: {
-    // loader: 'custom',
-    // loaderFile: 'src/lib/loader.js',
+    loader: 'custom',
+    loaderFile: 'src/lib/loader.js',
     remotePatterns: [
       {
         protocol: 'https',
